scripts/requirements.txt
# GEMINI API関連の依存パッケージ
google-generativeai>=0.4.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
flask>=3.0.0
flask-cors>=4.0.0

# PDF→画像変換（必要に応じて）
# PyMuPDF>=1.23.0
